+++ packages/playground/src/App.js
@@ -8,7 +8,7 @@ import ScrollToEnd, { StateContext } from 'react-scroll-to-bottom';
 import CommandBar from './CommandBar';
 
 const FADE_IN_ANIMATION = css.keyframes({
-  '0%': { opacity: .2 },
+  '0%': { opacity: 0.2 },
   '100%': { opacity: 1 }
 });
 
@@ -62,7 +62,7 @@ const SCROLL_VIEW_PADDING_CSS = css({
   },
 
   '& > p': {
-    animation: `${ FADE_IN_ANIMATION } 500ms`
+    animation: `${FADE_IN_ANIMATION} 500ms`
   }
 });
 
@@ -78,44 +78,69 @@ const App = () => {
   const [paragraphs, setParagraphs] = useState(createParagraphs(10));
   const [commandBarVisible, setCommandBarVisible] = useState(false);
 
-  const handleAdd = useCallback(count => setParagraphs([...paragraphs, ...createParagraphs(count)]), [paragraphs, setParagraphs]);
+  const handleAdd = useCallback(count => setParagraphs([...paragraphs, ...createParagraphs(count)]), [
+    paragraphs,
+    setParagraphs
+  ]);
   const handleAdd1 = useCallback(() => handleAdd(1), [handleAdd]);
   const handleAdd10 = useCallback(() => handleAdd(10), [handleAdd]);
+  const handleAddButton = useCallback(
+    () => setParagraphs([...paragraphs, 'Button: ' + loremIpsum({ units: 'words' })]),
+    [paragraphs, setParagraphs]
+  );
   const handleClear = useCallback(() => setParagraphs([]), [setParagraphs]);
-  const handleCommandBarVisibleChange = useCallback(({ target: { checked } }) => setCommandBarVisible(checked), [setCommandBarVisible]);
+  const handleCommandBarVisibleChange = useCallback(({ target: { checked } }) => setCommandBarVisible(checked), [
+    setCommandBarVisible
+  ]);
   const handleContainerSizeLarge = useCallback(() => setContainerSize('large'), [setContainerSize]);
   const handleContainerSizeNormal = useCallback(() => setContainerSize(''), [setContainerSize]);
   const handleContainerSizeSmall = useCallback(() => setContainerSize('small'), [setContainerSize]);
-  const handleIntervalEnabledChange = useCallback(({ target: { checked: intervalEnabled } }) => setIntervalEnabled(intervalEnabled), []);
-  const containerClassName = useMemo(() => classNames(
-    CONTAINER_CSS + '',
-    containerSize === 'small' ?
-      SMALL_CONTAINER_CSS + ''
-    : containerSize === 'large' ?
-      LARGE_CONTAINER_CSS + ''
-    :
-      ''
-  ), [containerSize]);
+  const handleIntervalEnabledChange = useCallback(
+    ({ target: { checked: intervalEnabled } }) => setIntervalEnabled(intervalEnabled),
+    []
+  );
+  const containerClassName = useMemo(
+    () =>
+      classNames(
+        CONTAINER_CSS + '',
+        containerSize === 'small' ? SMALL_CONTAINER_CSS + '' : containerSize === 'large' ? LARGE_CONTAINER_CSS + '' : ''
+      ),
+    [containerSize]
+  );
 
-  const handleKeyDown = useCallback(({ keyCode }) => {
-    switch (keyCode) {
-      case 49: return handleAdd1();
-      case 50: return handleAdd10();
-      case 51: return handleClear();
-      case 52: return handleContainerSizeSmall();
-      case 53: return handleContainerSizeNormal();
-      case 54: return handleContainerSizeLarge();
-      case 82: return window.location.reload(); // Press R key
-      default: break;
-    }
-  }, [
-    handleAdd1,
-    handleAdd10,
-    handleClear,
-    handleContainerSizeLarge,
-    handleContainerSizeNormal,
-    handleContainerSizeSmall
-  ]);
+  const handleKeyDown = useCallback(
+    ({ keyCode }) => {
+      switch (keyCode) {
+        case 49:
+          return handleAdd1();
+        case 50:
+          return handleAdd10();
+        case 51:
+          return handleClear();
+        case 52:
+          return handleContainerSizeSmall();
+        case 53:
+          return handleContainerSizeNormal();
+        case 54:
+          return handleContainerSizeLarge();
+        case 55:
+          return handleAddButton();
+        case 82:
+          return window.location.reload(); // Press R key
+        default:
+          break;
+      }
+    },
+    [
+      handleAdd1,
+      handleAdd10,
+      handleAddButton,
+      handleClear,
+      handleContainerSizeLarge,
+      handleContainerSizeNormal,
+      handleContainerSizeSmall
+    ]
+  );
 
   useEffect(() => {
     window.addEventListener('keydown', handleKeyDown);
@@ -124,7 +149,7 @@ const App = () => {
   }, [handleKeyDown]);
 
   return (
-    <div className={ROOT_CSS + ""}>
+    <div className={ROOT_CSS + ''}>
       <ul className="button-bar">
         <li>
           <button onClick={handleAdd1}>Add new paragraph</button>
@@ -145,39 +170,35 @@ const App = () => {
           <button onClick={handleContainerSizeLarge}>Large</button>
         </li>
         <li>
+          <button onClick={handleAddButton}>Add a button</button>
+        </li>
+        <li>
           <label>
-            <input
-              checked={intervalEnabled}
-              onChange={handleIntervalEnabledChange}
-              type="checkbox"
-            />
+            <input checked={intervalEnabled} onChange={handleIntervalEnabledChange} type="checkbox" />
             Add one every second
           </label>
         </li>
         <li>
           <label>
-            <input
-              checked={commandBarVisible}
-              onChange={handleCommandBarVisibleChange}
-              type="checkbox"
-            />
+            <input checked={commandBarVisible} onChange={handleCommandBarVisibleChange} type="checkbox" />
             Show command bar
           </label>
         </li>
       </ul>
       <div className="panes">
-        <ScrollToEnd
-          className={containerClassName}
-          scrollViewClassName={SCROLL_VIEW_CSS + ""}
-        >
+        <ScrollToEnd className={containerClassName} scrollViewClassName={SCROLL_VIEW_CSS + ''}>
           {commandBarVisible && <CommandBar />}
           <StateContext.Consumer>
             {({ sticky }) => (
-              <div
-                className={classNames(SCROLL_VIEW_PADDING_CSS + "", { sticky })}
-              >
+              <div className={classNames(SCROLL_VIEW_PADDING_CSS + '', { sticky })}>
                 {paragraphs.map(paragraph => (
-                  <p key={paragraph}>{paragraph}</p>
+                  <p key={paragraph}>
+                    {paragraph.startsWith('Button: ') ? (
+                      <button type="button">{paragraph.substr(8)}</button>
+                    ) : (
+                      paragraph
+                    )}
+                  </p>
                 ))}
               </div>
             )}
@@ -188,11 +209,15 @@ const App = () => {
           {commandBarVisible && <CommandBar />}
           <StateContext.Consumer>
             {({ sticky }) => (
-              <div
-                className={classNames(SCROLL_VIEW_PADDING_CSS + "", { sticky })}
-              >
+              <div className={classNames(SCROLL_VIEW_PADDING_CSS + '', { sticky })}>
                 {[...paragraphs].reverse().map(paragraph => (
-                  <p key={paragraph}>{paragraph}</p>
+                  <p key={paragraph}>
+                    {paragraph.startsWith('Button: ') ? (
+                      <button type="button">{paragraph.substr(8)}</button>
+                    ) : (
+                      paragraph
+                    )}
+                  </p>
                 ))}
               </div>
             )}
@@ -200,9 +225,7 @@ const App = () => {
           {commandBarVisible && <CommandBar />}
         </ScrollToEnd>
       </div>
-      {intervalEnabled && (
-        <Interval callback={handleAdd1} enabled={true} timeout={1000} />
-      )}
+      {intervalEnabled && <Interval callback={handleAdd1} enabled={true} timeout={1000} />}
     </div>
   );
 };